requirements.txt
aiofiles>=23.2.1
altair>=5.2.0
annotated-types>=0.6.0
anyio>=4.2.0
attrs>=23.2.0
baidu-aip>=4.16.13 # 使用BaiduASR时需要该库
certifi>=2023.11.17
chardet>=5.0.0
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
contourpy>=1.2.0
cycler>=0.12.1
distro>=1.9.0
fastapi>=0.109.0
ffmpy>=0.3.1
filelock>=3.13.1
fonttools>=4.47.2
fsspec>=2023.12.2
gradio>=4.14.0
gradio_client>=0.8.0
h11>=0.14.0
httpcore>=1.0.2
httpx>=0.26.0
huggingface-hub>=0.20.2
idna>=3.6
importlib-resources>=6.1.1
Jinja2>=3.1.3
jsonschema>=4.21.0
jsonschema-specifications>=2023.12.1
kiwisolver>=1.4.5
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
matplotlib>=3.8.2
mdurl>=0.1.2
numpy>=1.26.3
openai>=1.8.0 # 使用OpenAI的任意API时均需要该库
orjson>=3.9.15
packaging>=23.2
pandas>=2.1.4
pillow>=10.2.0
pydantic>=2.5.3
pydantic_core>=2.14.6
pydub>=0.25.1
Pygments>=2.17.2
pyparsing>=3.1.1
python-dateutil>=2.8.2
python-multipart>=0.0.6
pytz>=2023.3.post1
PyYAML>=6.0.1
referencing>=0.32.1
requests>=2.31.0
rich>=13.7.0
rpds-py>=0.17.1
scipy>=1.11.4
semantic-version>=2.10.0
shellingham>=1.5.4
six>=1.16.0
sniffio>=1.3.0
socksio>=1.0.0
starlette>=0.35.1
tomlkit>=0.12.0
toolz>=0.12.0
tqdm>=4.66.1
typer>=0.9.0
typing_extensions>=4.9.0
tzdata>=2023.4
urllib3>=2.1.0
uvicorn>=0.26.0
websockets>=11.0.3
